* master:
  Select lua context per thread (#7465)
  Fix out of bounds access error in jtest (#7526)
  Disable compiling Inline.cc on macOS (#7389)
  Makes sure the types are correct, avoiding compiler warnings (#7523)
  Move has_request_body to ProxyTransaction (#7499)
  Make the H3 build script work properly on Debian platforms (#7522)
  slice/handleFirstServerHeader: return sooner on requested range errors (#7486)
  Add new log field for negotiated ALPN Protocol ID with the client (#7491)
  Add Outbound PROXY Protocol (v1/v2) Support (#7446)
  Updates the Dockerfile for debian (#7518)
  Disable client inactivity timeout while server is processing POST request (#7309)
  Upgrade Catch.hpp to v2.13.4 (#7464)
  Move reopen_moved_log_files to log flushing thread (#7450)
  replace psutil.pid() with psutil.process_iter() for safer execution (#7515)
  Fix spacing in clang-analyzer.sh script (#7480)
  Fix out of bounds access error in ats_base64_decode (#7490)
  Updated to build lastest versions of Fedora and CentOS docker images (#7505)
  Fix QUIC unit tests build issue on GNU ld (#7496)
  Fix QUIC unit test failures (#7497)
  Fixed build issues with Fedora 34 (#7506)
  Fixing DNS local_ipv* config option (#7507)
  traffic_dump: AuTests to use Proxy Verifier. (#7502)
  Disable ja3 plugin when building with boringssl (#7500)
  Avoid -Warray-bounds on PROXY Protocol Builder (#7488)
  AuTest: Upgrade to Proxy Verifier 2.0.2 (#7493)
  fix certs (#7494)
  Add zlib1g-dev to Debian dependencies in README (#7495)
  Unit Test -  Increase openssl's key size. Place test certs into a common test folder. (#7451)
  Add basic type aliases for std::chrono types to ink_time.h for future use. (#7482)
  traffic_ctl - Fix lookup key for run-root option (#7484)
  update thread config tests (#7370)
  Perf: Replace casecmp with memcmp in HPACK static table lookup (#6521)
  Add PROXY Protocol Builder (#7445)
  Adjust so transfer-encoding header can be treated hop-by-hop (#7473)
  Convert auxkey form 2 uint32_t to 1 uint64_t. (#7350)
  Remove the queuing option from proxy.config.http.per_server.connection (#7302)
  Remove unused function ink_microseconds. (#7481)
  use std::unordered_map to store sessions (#7405)
  drop use of BIO_f_base64 and EVP_PKEY_new_mac_key (#7106)
  Do not write to the cache if the plugin decides not to write to the cache (#7461)
  API to retrieve NoStore set by plugins (#7439)
  Update AuTest version update directions for pipenv (#7469)
  Add command line utility to help convert remap plugin usage to ATS9. (#7426)
  Cleanup: Get rid of MIMEFieldWrapper from HPACK encoding (#6520)
  Proxy Verifier: Making use of delay directives for caching tests. (#7468)
  Cleanup: Add SNIRoutingType (#7453)
  Updating to Proxy Verifier v2.0.0 (#7454)
  Adjust to actually try a server address more than once (#7288)
  Change atoi to atol, causing obvious issues on what needs to be int64's (#7466)
  Cleans up duplicated TSOutboundConnectionMatchType definition (#7090)
  Fixing compress expectation for new microserver (#7463)
  Update to the new MicroServer 1.0.6 release (#7460)
  CacheRead: clear dir entry if doc is found to be truncated (#7064)
  Do not provide a stale negative cache (#7422)
  Generalize SNI support (#6870)
  Add synchronization between UDPNetProcessor::UDPBind in main Thread and initialize_thread_for_udp_net in ET_UDP Thread (#7407)
  Fix heap use after free in DNSProcessor::getby() (#3871)
  Fix comment in include/tscore/Filenames.h. (#7457)
  Fix Makefile target for creating changelogs (#7455)
  Change squid log code for self looping (#7443)
  Enhancements for compress plugin (#7416)
  Add incoming PROXY Protocol v2 support (#7340)
  Cleanup: Remove unused members of NextHopProperty (#7436)
  Small fix to regex_remap PR # 7347. (#7437)
  PoolableSession (#6828)
  option to disable compression for range request's response (#7287)
  Make TSUrlSchemeGet() return scheme implied by URL type when there is no explicit scheme. (#7262)

+++ proxy/hdrs/unit_tests/test_HdrUtils.cc
@@ -160,7 +160,7 @@ TEST_CASE("HdrUtils 2", "[proxy][hdrutils]")
   int skip   = 0;
   auto parse = mime_hdr_print(heap, mime.m_mime, buff, static_cast<int>(sizeof(buff)), &idx, &skip);
   REQUIRE(parse != 0);
-  REQUIRE(idx == text.size());
+  REQUIRE(idx == static_cast<int>(text.size()));
   REQUIRE(0 == memcmp(ts::TextView(buff, idx), text));
   heap->destroy();
 };
@@ -203,7 +203,7 @@ TEST_CASE("HdrUtils 3", "[proxy][hdrutils]")
   int skip   = 0;
   auto parse = mime_hdr_print(heap, mime.m_mime, buff, static_cast<int>(sizeof(buff)), &idx, &skip);
   REQUIRE(parse != 0);
-  REQUIRE(idx == text.size());
+  REQUIRE(idx == static_cast<int>(text.size()));
   REQUIRE(0 == memcmp(ts::TextView(buff, idx), text));
   heap->destroy();
 };
